Break out get api function
