+++ homeassistant/components/unifi/hub/__init__.py
@@ -1,3 +1,4 @@
 """Internal functionality not part of HA infrastructure."""
 
-from .hub import UnifiHub, get_unifi_api  # noqa: F401
+from .api import get_unifi_api  # noqa: F401
+from .hub import UnifiHub  # noqa: F401
